requirements.txt
# MCP Server dependencies
mcp>=1.1.0

# Perplexity API client dependencies
httpx>=0.27.1
python-dotenv==1.0.0
